requirements.txt
argon2-cffi==19.1.0
asgi-redis==1.4.3
asgiref==3.1.4
bcrypt==3.1.7
channels-redis==2.4.0
Django==2.2.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
Pillow==6.1.0
pkg-resources==0.0.0
requests==2.22.0
wheel==0.33.4
